+++ client/client_test.go
@@ -1118,7 +1118,11 @@ func TestClient_UpdateNodeFromDevicesAccumulates(t *testing.T) {
 		Disk:         client.configCopy.Node.NodeResources.Disk,
 
 		// injected
-		Cpu:    structs.NodeCpuResources{CpuShares: 123},
+		Cpu: structs.NodeCpuResources{
+			CpuShares:          123,
+			ReservableCpuCores: client.configCopy.Node.NodeResources.Cpu.ReservableCpuCores,
+			TotalCpuCores:      client.configCopy.Node.NodeResources.Cpu.TotalCpuCores,
+		},
 		Memory: structs.NodeMemoryResources{MemoryMB: 1024},
 		Devices: []*structs.NodeDeviceResource{
 			{
@@ -1156,7 +1160,11 @@ func TestClient_UpdateNodeFromDevicesAccumulates(t *testing.T) {
 		Disk:         client.configCopy.Node.NodeResources.Disk,
 
 		// injected
-		Cpu:    structs.NodeCpuResources{CpuShares: 123},
+		Cpu: structs.NodeCpuResources{
+			CpuShares:          123,
+			ReservableCpuCores: client.configCopy.Node.NodeResources.Cpu.ReservableCpuCores,
+			TotalCpuCores:      client.configCopy.Node.NodeResources.Cpu.TotalCpuCores,
+		},
 		Memory: structs.NodeMemoryResources{MemoryMB: 2048},
 		Devices: []*structs.NodeDeviceResource{
 			{
@@ -1178,6 +1186,9 @@ func TestClient_UpdateNodeFromDevicesAccumulates(t *testing.T) {
 // network interfaces take precedence over fingerprinted ones.
 func TestClient_UpdateNodeFromFingerprintKeepsConfig(t *testing.T) {
 	t.Parallel()
+	if runtime.GOOS != "linux" {
+		t.Skip("assertions assume linux platform")
+	}
 
 	// Client without network configured updates to match fingerprint
 	client, cleanup := TestClient(t, nil)