+++ client/fingerprint/cpu_default.go
@@ -1,0 +1,7 @@
+//+build !linux
+
+package fingerprint
+
+func (f *CPUFingerprint) deriveReservableCores(req *FingerprintRequest) ([]uint16, error) {
+	return nil, nil
+}