+++ client/lib/cgutil/cgutil_default.go
@@ -1,0 +1,13 @@
+// +build !linux
+
+package cgutil
+
+const (
+	DefaultCgroupParent = ""
+)
+
+// FindCgroupMountpointDir is used to find the cgroup mount point on a Linux
+// system. Here it is a no-op implemtation
+func FindCgroupMountpointDir() (string, error) {
+	return "", nil
+}