+++ e2e/connect/client.go
@@ -9,9 +9,13 @@ import (
 	"github.com/hashicorp/nomad/helper/uuid"
 )
 
+const (
+	consulNamespace = "default"
+)
+
 type ConnectClientStateE2ETest struct {
 	framework.TC
-	jobIds []string
+	jobIDs []string
 }
 
 func (tc *ConnectClientStateE2ETest) BeforeAll(f *framework.F) {
@@ -24,10 +28,10 @@ func (tc *ConnectClientStateE2ETest) AfterEach(f *framework.F) {
 		return
 	}
 
-	for _, id := range tc.jobIds {
+	for _, id := range tc.jobIDs {
 		tc.Nomad().Jobs().Deregister(id, true, nil)
 	}
-	tc.jobIds = []string{}
+	tc.jobIDs = []string{}
 	tc.Nomad().System().GarbageCollect()
 }
 
@@ -35,7 +39,7 @@ func (tc *ConnectClientStateE2ETest) TestClientRestart(f *framework.F) {
 	t := f.T()
 
 	jobID := "connect" + uuid.Generate()[0:8]
-	tc.jobIds = append(tc.jobIds, jobID)
+	tc.jobIDs = append(tc.jobIDs, jobID)
 	client := tc.Nomad()
 	consulClient := tc.Consul()
 
@@ -43,18 +47,16 @@ func (tc *ConnectClientStateE2ETest) TestClientRestart(f *framework.F) {
 		"connect/input/demo.nomad", jobID, "")
 	f.Equal(2, len(allocs))
 
-	e2eutil.RequireConsulStatus(f.Assertions, consulClient,
-		"count-api-sidecar-proxy", capi.HealthPassing)
+	e2eutil.RequireConsulStatus(f.Assertions, consulClient, consulNamespace, "count-api-sidecar-proxy", capi.HealthPassing)
 	nodeID := allocs[0].NodeID
 
 	restartID, err := e2eutil.AgentRestart(client, nodeID)
 	if restartID != "" {
-		tc.jobIds = append(tc.jobIds, restartID)
+		tc.jobIDs = append(tc.jobIDs, restartID)
 	}
 	if err != nil {
 		t.Skip("node cannot be restarted", err)
 	}
 
-	e2eutil.RequireConsulStatus(f.Assertions, consulClient,
-		"count-api-sidecar-proxy", capi.HealthPassing)
+	e2eutil.RequireConsulStatus(f.Assertions, consulClient, consulNamespace, "count-api-sidecar-proxy", capi.HealthPassing)
 }