+++ plugins/drivers/utils.go
@@ -142,8 +142,8 @@ func ResourcesFromProto(pb *proto.Resources) *Resources {
 			CPUShares:        pb.LinuxResources.CpuShares,
 			MemoryLimitBytes: pb.LinuxResources.MemoryLimitBytes,
 			OOMScoreAdj:      pb.LinuxResources.OomScoreAdj,
-			CpusetCPUs:       pb.LinuxResources.CpusetCpus,
-			CpusetMems:       pb.LinuxResources.CpusetMems,
+			CpusetCpus:       pb.LinuxResources.CpusetCpus,
+			CpusetCgroupPath: pb.LinuxResources.CpusetCgroup,
 			PercentTicks:     pb.LinuxResources.PercentTicks,
 		}
 	}
@@ -212,8 +212,8 @@ func ResourcesToProto(r *Resources) *proto.Resources {
 			CpuShares:        r.LinuxResources.CPUShares,
 			MemoryLimitBytes: r.LinuxResources.MemoryLimitBytes,
 			OomScoreAdj:      r.LinuxResources.OOMScoreAdj,
-			CpusetCpus:       r.LinuxResources.CpusetCPUs,
-			CpusetMems:       r.LinuxResources.CpusetMems,
+			CpusetCpus:       r.LinuxResources.CpusetCpus,
+			CpusetCgroup:     r.LinuxResources.CpusetCgroupPath,
 			PercentTicks:     r.LinuxResources.PercentTicks,
 		}
 	}